Changed framebuffer width, height default to None.
I changed the framebuffer width and height default value to none, and corrected the checks in the color, depth and stencil setters accordingly. The reason None is better is because it is possible to have a renderbuffer with e.g. 0 width and 10 height.
Compare np.zeros((0, 10)), which is allowed in numpy. According to the opengl documentation the only limitation is that it should be nonnegative.